pyproject.toml
[project]
name = "clipped"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.7",
    "huggingface-hub>=0.27.1",
    "onnxruntime>=1.20.1",
    "optimum>=1.23.3",
    "pillow>=11.1.0",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "timm>=1.0.14",
    "torch>=2.6.0.dev20241213",
    "torchvision>=0.22.0.dev20241213",
    "transformers>=4.48.1",
    "uvicorn>=0.34.0",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/nightly/cpu"
explicit = true
